The MATRIX 1984 Beta v0.14
I rushed this fix out because the game is searching for a saved game file and because I have had one on my machine, I didn't notice the actual bug was that new players wouldn't have that. I've made a few changes and this one is working for me. Thanks to Jared Keown for reporting this bug.